+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 15/03/2021
+- Mise en place du sel
+- Mise en place de l'auxilaire 2
+
 # 28/12/2021
 - Mise en place du graphe pour la 4.2 sur la tuile
 - Optimisation du code JS en Vanilla